Targeting changed to only include .NET 6.0
