+++ Frends.PowerShell.RunScript/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+# [1.0.1] - 2022-03-02
+### Added
+- Targeting changed to only include .NET 6.0 (removed .NET Standard 2.0)
+
 ## [1.0.0] - 2022-02-25
 ### Added
 - Initial implementation
